Add Change Coupling Metrics functionality
The update introduces Change Coupling metrics that help to represent the change coupling between two files in the same repository. It includes a record to store change coupling instances, a CSV-based report for the change coupling, and updates to the Analysis RunCommand. The README is also updated to reflect this new feature.

https://github.com/bananen-software/gavel/issues/10