Add logging and configurable analysis timeframe in RunAnalysisCommand
The option to set the timeframe of analysis has been added to the GitConfig in the config.json file, which impacts the RunAnalysisCommand to pull commits. Concurrently, logging capabilities have been added to RunAnalysisCommand to provide more detailed information during the analysis process, such as the number of mailmap entries loaded and the number of commits found within the set timeframe.