Add relational cohesion status in metrics table
A new javascript file, relational-cohesion.js has been added to determine the relational cohesion status. An additional 'Status' column has been added in the HTML to display the determined status in the relational-cohesion-metrics table. The components' cohesion status is calculated and displayed as a percentage along with the other metrics.